requirements.txt
absl-py==2.0.0
aiohttp==3.9.0
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.1.0
awkward==2.5.0
awkward-cpp==26
awkward0==0.15.5
blosc2==2.3.1
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.40
glob2==0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning==2.1.2
lightning-utilities==0.10.0
lz4==4.3.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
ndindex==1.7
networkx==3.0
numexpr==2.8.7
numpy==1.24.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
plotly==5.18.0
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sentry-sdk==1.35.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tables==3.9.1
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
torch==2.1.1
torchaudio==2.1.1
torchmetrics==1.2.0
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.4.0
tzdata==2023.3
uproot==5.1.2
urllib3==1.26.13
vector==1.1.1.post1
wandb==0.16.0
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
